Feat: Add user guard

+++ backend/dependencies/CommentReaction/CommentReactionController.ts
@@ -15,7 +15,10 @@ export class CommentReactionController {
   async createReaction(req: Request, res: Response, next: NextFunction) {
     try {
       const reactionBody = req.body.data;
-      const user: currentUser = req.currentUser;
+      const user: currentUser | undefined = req.currentUser;
+      if (!user) {
+        throw new ErrorHandler('Unauthorised', 401);
+      }
       const payload: CommentReactionInsertDTO = {
         author_id: user.id,
         type: reactionBody.type,
@@ -51,7 +54,10 @@ export class CommentReactionController {
   }
   async deleteReaction(req: Request, res: Response, next: NextFunction) {
     const reactionBody = req.body.data;
-    const user: currentUser = req.currentUser;
+    const user: currentUser | undefined = req.currentUser;
+    if (!user) {
+      throw new ErrorHandler('Unauthorised', 401);
+    }
     try {
       const result = await this.commentReactionService.deleteUserReactionInComment(user.id, reactionBody.commentId);
       if (!result) {
